# portlist: document, clean up, fix an open fd spike, optimize a bit