--- a/portlist/portlist_other.go
+++ b/portlist/portlist_other.go
@@ -6,8 +6,10 @@
 
 package portlist
 
+import "time"
+
 // We have to run netstat, which is a bit expensive, so don't do it too often.
-const POLL_SECONDS = 5
+const pollInterval = 5 * time.Second
 
 func listPorts() (List, error) {
 	return listPortsNetstat("-na")