@danderson 2020-03-14
___
*[ipn/local.go, line 101 at r1](https://reviewable.io/reviews/tailscale/tailscale/186#-M2MDjmH7TVexLs8CU7v:-M2MDjmH7TVexLs8CU7w:bb0ujyy) ([raw file](https://github.com/tailscale/tailscale/blob/f109a11189598e5d445073e062849b68575625e1/ipn/local.go#L101)):*
> ```Go
> 
> func (b *LocalBackend) Shutdown() {
> 	b.ctxCancel()
> ```

Any shutdown races? Do we need to block until the poller goroutine shuts down to avoid problems?

(spidey sense tripping because I've fixed 2 or 3 such races in magicsock recently)


<!-- Sent from Reviewable.io -->